moving db models to subfolder
